Refactor tests for reply deletion
- refactor reply delete request for status 200
- add reply delete request for status 404
- indicate missing tests for status 400, 401 and 402
- drop old implementation
- refactor comment delete tests to match reply delete tests

See #241
